Fix Texts.tbl 4523 and 4552 format specifiers.
Note that it was for sell and buy strings, where the format specifier
supposed to be an integer (%d), but it was there as a string (%s).
The way they handled it in the official client, is by casting it to
string and then to c string to format the buffer lol...
Instead the mistake was in the tbl, since there are no more
usages for these ids except for buy and sell ui-tooltips.