Revert "Fix Texts.tbl 4523 and 4552 format specifiers."
This reverts commit 1809e3c0e994c0f5166ebcc3b50b64f192b962ed.

I was stupid and under the wrong assumption. The client was at fault,
not the tbl. The client provided it an int instead of a string. That
string is the coins, but formatted with commas. Therefore I had to
implement that function to format coins with commas in the client before
using this tbl keys for formatting the string.